Relay flow types

+++ app/fbcnms-projects/inventory/app/components/configure/LocationTypeItem.js
@@ -98,6 +98,7 @@ class LocationTypeItem extends React.Component<Props> {
               <div className={classes.properties}>
                 <DynamicPropertyTypesGrid
                   key={locationType.id}
+                  // $FlowFixMe[incompatible-type] $FlowFixMe T74239404 Found via relay types
                   propertyTypes={locationType.propertyTypes}
                 />
               </div>
@@ -125,18 +126,17 @@ class LocationTypeItem extends React.Component<Props> {
             },
           },
           store => {
-            // $FlowFixMe (T62907961) Relay flow types
             const rootQuery = store.getRoot();
             const locationTypes = ConnectionHandler.getConnection(
               rootQuery,
               'Catalog_locationTypes',
             );
-            ConnectionHandler.deleteNode(
-              // $FlowFixMe (T62907961) Relay flow types
-              locationTypes,
-              this.props.locationType.id,
-            );
-            // $FlowFixMe (T62907961) Relay flow types
+            if (locationTypes != null) {
+              ConnectionHandler.deleteNode(
+                locationTypes,
+                this.props.locationType.id,
+              );
+            }
             store.delete(this.props.locationType.id);
           },
         );
